Fixed Dirac Delta distribution/savings accts

+++ MonteCarlo/Models/Result.cs
@@ -4,7 +4,8 @@ namespace MonteCarlo.Models
 {
     public class Result
     {
-        public double SuccessRate { get; set; }
+        public int SuccessRate { get; set; }
         public Dictionary<string, double[][]> Trials { get; set; }
+        //public PortfolioPercentiles;
     }
 }
